Add per route Middleware
This allows each route to configure the Middleware that runs prior to
the actual handler. This gives each route the ability to
enable/disable/configure common Middleware functions.

This replaces the global middleware and forces each route to define the
set of middleware (or inherit the default). The ordering of middleware
functions is extremely important and potentially error prone, so best
not to have 2 levels of middleware functions.

Additionally this installs the source.Middleware function for all query
endpoints. This made it possible to change the semantics of the function
to handle the M3-Source header missing, instead of skipping the
function.
