+++ src/query/source/source_test.go
@@ -111,6 +111,7 @@ func TestMiddleware(t *testing.T) {
 		name          string
 		sourceHeaders []string
 		expected      testSource
+		expectedLog   string
 		deserializer  Deserializer
 		invalidErr    bool
 	}{
@@ -118,11 +119,16 @@ func TestMiddleware(t *testing.T) {
 			name:          "happy path",
 			sourceHeaders: []string{"foobar"},
 			expected:      testSource{"foobar"},
+			expectedLog:   "foobar",
 		},
 		{
 			name:          "no source header",
 			sourceHeaders: []string{},
-			expected:      testSource{},
+			expected:      testSource{"nil"},
+			deserializer: func(bytes []byte) (interface{}, error) {
+				require.Nil(t, bytes)
+				return testSource{name: "nil"}, nil
+			},
 		},
 		{
 			name:          "multiple source headers",
@@ -182,9 +188,9 @@ func TestMiddleware(t *testing.T) {
 				entry := testMsgs[0]
 				require.Equal(t, "test", entry.Message)
 				fields := entry.ContextMap()
-				if tc.expected.name != "" {
+				if tc.expectedLog != "" {
 					require.Len(t, fields, 1)
-					require.Equal(t, tc.expected.name, fields["source"])
+					require.Equal(t, tc.expectedLog, fields["source"])
 				}
 			}
 		})
